requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
qrcode[pil]==7.4.2
Pillow==10.0.1